+++ packages/builder/builder-rspack-provider/src/plugins/resolve.ts
@@ -10,7 +10,7 @@ export const builderPluginResolve = (): BuilderPlugin => ({
   setup(api) {
     applyBuilderResolvePlugin(api);
 
-    api.modifyRspackConfig(async rspackConfig => {
+    api.modifyRspackConfig(async (rspackConfig, { isServer }) => {
       const isTsProject = Boolean(api.context.tsconfigPath);
 
       if (isTsProject) {
@@ -19,6 +19,15 @@ export const builderPluginResolve = (): BuilderPlugin => ({
           'resolve.tsConfigPath',
           api.context.tsconfigPath,
         );
+      }
+
+      if (isServer) {
+        // FIXME:
+        // When targe = node, we no need to specify conditionsNames.
+        // We gueess the webpack would auto specify refrence to target.
+        // Rspack has't the action, so we need manually specify.
+        const nodeConditionNames = ['require', 'node'];
+        setConfig(rspackConfig, 'resolve.conditionNames', nodeConditionNames);
       }
     });
   },