+++ packages/builder/builder-shared/src/types/config/tools.ts
@@ -1,6 +1,10 @@
-import type { ChainedConfig } from '../utils';
+import type { ChainedConfig, FileFilterUtil } from '../utils';
 import type { DevServerOptions } from '@modern-js/types';
-import type { AutoprefixerOptions } from '../thirdParty';
+import type {
+  AutoprefixerOptions,
+  SassLoaderOptions,
+  LessLoaderOptions,
+} from '../thirdParty';
 
 /** html-rspack-plugin is compatible with html-webpack-plugin */
 export type { Options as HTMLPluginOptions } from 'html-webpack-plugin';
@@ -8,6 +12,16 @@ export type { Options as HTMLPluginOptions } from 'html-webpack-plugin';
 export type ToolsDevServerConfig = ChainedConfig<DevServerOptions>;
 
 export type ToolsAutoprefixerConfig = ChainedConfig<AutoprefixerOptions>;
+
+export type ToolsSassConfig = ChainedConfig<
+  SassLoaderOptions,
+  { addExcludes: FileFilterUtil }
+>;
+
+export type ToolsLessConfig = ChainedConfig<
+  LessLoaderOptions,
+  { addExcludes: FileFilterUtil }
+>;
 
 export interface SharedToolsConfig {
   /**