+++ packages/builder/builder-webpack-provider/src/types/thirdParty/index.ts
@@ -1,13 +1,10 @@
 import type WebpackChain from '@modern-js/builder-shared/webpack-5-chain';
 import type webpack from 'webpack';
-import type { Configuration as WebpackConfig, LoaderContext } from 'webpack';
+import type { Configuration as WebpackConfig } from 'webpack';
 import type TerserPlugin from 'terser-webpack-plugin';
 import type CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
 import type ForkTSCheckerPlugin from 'fork-ts-checker-webpack-plugin';
 import type { Options as RawTSLoaderOptions } from 'ts-loader';
-import type { Options as SassOptions } from '../../../compiled/sass';
-import type * as SassLoader from '../../../compiled/sass-loader';
-import type Less from '../../../compiled/less';
 
 export type { BabelTransformOptions } from '@modern-js/types';
 export type { InspectorPluginOptions } from '@modern-js/inspector-webpack-plugin';
@@ -43,20 +40,3 @@ export type {
   MiniCSSExtractPluginOptions,
   MiniCSSExtractLoaderOptions,
 } from './css';
-
-export type SassLoaderOptions = Omit<SassLoader.Options, 'sassOptions'> & {
-  sassOptions?: SassOptions<'sync'>;
-};
-
-export type LessLoaderOptions = {
-  lessOptions?: Less.Options;
-  additionalData?:
-    | string
-    | ((
-        content: string,
-        loaderContext: LoaderContext<LessLoaderOptions>,
-      ) => string);
-  sourceMap?: boolean;
-  webpackImporter?: boolean;
-  implementation?: unknown;
-};